B"H In this adjustment:
We're adding minimapCamera.right and minimapCamera.top to the player's coordinates within the minimap. This is to account for the fact that the minimap camera's position is not always at (0, maxY, 0) but could be anywhere within the scene.
Then we divide by the width and height of the minimap's viewport to get the relative position within the minimap.
The rest of the calculation remains the same as before to translate this relative position into scroll positions for the parent element.